requirements.txt
Flask
yfinance
pandas
plotly
newsapi-python
python-dotenv